@@ -25,29 +25,31 @@
   },
   "homepage": "https://github.com/reportportal/documentation#readme",
   "dependencies": {
-    "@docusaurus/core": "^2.4.0",
-    "@docusaurus/preset-classic": "^2.4.0",
-    "@mdx-js/react": "^1.6.22",
-    "clsx": "^1.2.1",
-    "dotenv": "^16.0.3",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-image-video-lightbox": "^3.0.1",
-    "prop-types": "^15.8.1"
+    "@docusaurus/core": "3.0.1",
+    "@docusaurus/preset-classic": "3.0.1",
+    "@mdx-js/react": "3.0.0",
+    "prism-react-renderer": "^2.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "clsx": "^2.0.0",
+    "dotenv": "^16.3.1",
+    "prop-types": "^15.8.1",
+    "yet-another-react-lightbox": "^3.15.6"
   },
   "devDependencies": {
-    "@docusaurus/eslint-plugin": "^2.4.0",
-    "@docusaurus/module-type-aliases": "^2.4.0",
-    "eslint": "^8.37.0",
+    "@docusaurus/eslint-plugin": "^3.0.1",
+    "@docusaurus/types": "3.0.1",
+    "@docusaurus/module-type-aliases": "^3.0.1",
+    "docusaurus-mdx-checker": "^3.0.0",
+    "eslint": "^8.56.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.32.2",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "^5.1.0",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-import": "^2.27.5",
-    "prettier": "^2.8.7"
+    "prettier": "^3.1.1"
   },
   "browserslist": {
     "production": [
@@ -62,6 +64,6 @@
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
