@@ -16,13 +16,15 @@
 
 import BrowserOnly from '@docusaurus/BrowserOnly';
 import PropTypes from 'prop-types';
-import React, { useState } from 'react';
+import * as React from 'react';
+import { useState } from 'react';
+import Lightbox from 'yet-another-react-lightbox';
 
 import styles from './mediaViewer.module.css';
 
-const ReactImageVideoLightbox = require('react-image-video-lightbox').default;
+import 'yet-another-react-lightbox/styles.css';
 
-const TYPE_PHOTO = 'photo';
+const TYPE_PHOTO = 'image';
 const TYPE_VIDEO = 'video';
 
 const getSrc = (src) => {
@@ -39,34 +41,38 @@ export function MediaViewer({ src, type, alt, thumbnail }) {
       {() => {
         let contentSrc = getSrc(src);
         const thumbnailSrc = thumbnail ? getSrc(thumbnail) : contentSrc;
-        const isVideo = type === TYPE_VIDEO;
 
-        if (isVideo && contentSrc.includes('youtu.be/')) {
+        if (type === TYPE_VIDEO && contentSrc.includes('youtu.be/')) {
           contentSrc = contentSrc.replace('.be/', 'be.com/embed/');
+          return (
+            <iframe
+              className={styles['video-iframe']}
+              src={contentSrc}
+              title={alt}
+              allow="accelerometer; clipboard-write; encrypted-media; gyroscope; web-share"
+              allowFullScreen
+            />
+          );
         }
-
         return (
           <>
-            <button
-              type="button"
-              className={styles[isVideo ? 'video-container' : 'media-container']}
-              onClick={openViewer}
-              onKeyPress={openViewer}
-            >
+            <button type="button" className={styles['media-container']} onClick={openViewer}>
               <img className={styles.thumbnail} src={thumbnailSrc} alt={alt} />
             </button>
             {open && (
               <div className={styles['preview-container']}>
-                <ReactImageVideoLightbox
-                  data={[
-                    {
-                      url: contentSrc,
-                      type,
-                      alt,
+                <Lightbox
+                  slides={[{ type, src: contentSrc, alt }]}
+                  open={open}
+                  close={() => setOpen(false)}
+                  render={{
+                    iconPrev() {
+                      return null;
                     },
-                  ]}
-                  startIndex={0}
-                  onCloseCallback={() => setOpen(false)}
+                    iconNext() {
+                      return null;
+                    },
+                  }}
                 />
               </div>
             )}
