@@ -16,9 +16,8 @@
 
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { PageMetadata } from '@docusaurus/theme-common';
-import Layout from '@theme/Layout';
 import React from 'react';
-
+import Heading from '@theme/Heading';
 import styles from './notFoundBlock.module.css';
 import { SearchButton } from './searchButton';
 
@@ -26,23 +25,23 @@ export function NotFoundBlock() {
   return (
     <>
       <PageMetadata title="Page Not Found" />
-      <Layout>
-        <div className={styles['not-found-content']}>
-          <div className={styles['not-found-img']} />
-          <h1 className={styles['not-found-title']}>Page Not Found</h1>
-          <p className={styles['not-found-text']}>
-            You may have used an outdated link as there have been some changes in the structure of
-            the documentation.
-          </p>
-          <p className={styles['not-found-text']}>
-            But it/'s clearly here somewhere! <br />
-            Please, use the <strong>keyword search</strong> to find it!
-          </p>
-          <div className={styles['search-button-wrapper']}>
-            <SearchButton />
-          </div>
+      <div className={styles['not-found-content']}>
+        <div className={styles['not-found-img']} />
+        <Heading as="h1" className={styles['not-found-title']}>
+          Page Not Found
+        </Heading>
+        <p className={styles['not-found-text']}>
+          You may have used an outdated link as there have been some changes in the structure of
+          documentation.
+        </p>
+        <p className={styles['not-found-text']}>
+          But it/'s clearly here somewhere! <br />
+          Please, use the <strong>keyword search</strong> to find it!
+        </p>
+        <div className={styles['search-button-wrapper']}>
+          <SearchButton />
         </div>
-      </Layout>
+      </div>
     </>
   );
 }
