@@ -26,8 +26,7 @@ import { createPortal } from 'react-dom';
 import { DocSearchModal, useDocSearchKeyboardEvents } from '@docsearch/react';
 import Link from '@docusaurus/Link';
 import { useHistory } from '@docusaurus/router';
-import { isRegexpStringMatch } from '@docusaurus/theme-common';
-import { useSearchPage } from '@docusaurus/theme-common/internal';
+import { isRegexpStringMatch, useSearchLinkCreator } from '@docusaurus/theme-common';
 import {
   useAlgoliaContextualFacetFilters,
   useSearchResultUrlProcessor,
@@ -44,11 +43,13 @@ function Hit({ hit, children }) {
 }
 
 function ResultsFooter({ state, onClose }) {
-  const { generateSearchPageLink } = useSearchPage();
+  const createSearchLink = useSearchLinkCreator();
 
   return (
-    <Link to={generateSearchPageLink(state.query)} onClick={onClose}>
-      <Translate id="theme.SearchBar.seeAll" values={{ count: state.context.nbHits }}>
+    <Link to={createSearchLink(state.query)} onClick={onClose}>
+      <Translate
+        id="theme.SearchBar.seeAll"
+        values={{count: state.context.nbHits}}>
         {'See all {count} results'}
       </Translate>
     </Link>
